Convenience methods from `reply` now on ResponseWriter; Identity takes ResponseWriter
